+++ project/resources/car.vert
@@ -1,31 +1,31 @@
 #version 130
 
-uniform mediump mat4 ViewMatrix;
-uniform mediump mat4 ModelMatrix;
-uniform mediump mat4 ProjectionMatrix;
+uniform highp mat4 ViewMatrix;
+uniform highp mat4 ModelMatrix;
+uniform highp mat4 ProjectionMatrix;
 
-uniform mediump mat3 NormalMatrix;
+uniform highp mat3 NormalMatrix;
 
-uniform mediump vec4 LightPos;
-uniform mediump vec4 EyePos;
+uniform highp vec4 LightPos;
+uniform highp vec4 EyePos;
 
-uniform mediump float DayNightPulse;
-uniform mediump vec3 LightDirection;
+uniform highp float DayNightPulse;
+uniform highp vec3 LightDirection;
 
 attribute vec4 Position;
 attribute vec3 Normal;
 attribute vec3 Tangent;
 attribute vec4 TexCoord;
 
-varying lowp vec4 diffuseVarying;
-varying lowp vec4 texCoordVarying;
+varying highp vec4 diffuseVarying;
+varying highp vec4 texCoordVarying;
 
-varying mediump vec4 posVarying;        // pos in world space
-varying mediump vec3 normalVarying;     // normal in world space
-varying mediump vec3 tangentVarying;    // tangent in world space
+varying highp vec4 posVarying;        // pos in world space
+varying highp vec3 normalVarying;     // normal in world space
+varying highp vec3 tangentVarying;    // tangent in world space
 
-varying mediump float DayNightPulseVarying; // Value for the day and night cycle
-varying mediump vec3 LightDirectionVarying; // Value for the direction light
+varying highp float DayNightPulseVarying; // Value for the day and night cycle
+varying highp vec3 LightDirectionVarying; // Value for the direction light
 
 void main()
 {
